renamed files and added structure for error classes
