--- logsim/old/error_chen.py
@@ -1,17 +1,0 @@
-
-class ParserError(Exception):
-
-    def __init__(self, symbol, message): 
-        self.symbol = symbol 
-        self.message = message 
-    
-    def __eq__(self, error2): 
-        return type(self) == type(error2) and \
-            self.symbol == error2.symbol and \
-            self.message == error2.message
-
-class ParserSyntaxError(ParserError): 
-    pass 
-
-class ParserSyntaxError(ParserError): 
-    pass
